Move new-navbar-auth-dropdown to new-osf-navbar/auth-dropdown
